Renamed drug to resistance package
